TestTask.MudBlazor - Added implementation of the "IMessageBox" interface in the descendant class "DialogServiceShow" and Refactoring folder.
