+++ TestTask.MudBlazors/Dialog/DialogYesNo.razor
@@ -3,15 +3,18 @@
         @Content
     </DialogContent>
     <DialogActions>
-        <MudButton Variant="Variant.Filled" Class="ml-2" OnClick="Submit">Yes</MudButton>
-        <MudButton Variant="Variant.Filled" OnClick="Cancel">No</MudButton>
+        <MudButton Variant="Variant.Filled" Class="ml-2" OnClick="Submit">@SelectButtonDialog[true]</MudButton>
+        <MudButton Variant="Variant.Filled" OnClick="Cancel">@SelectButtonDialog[false]</MudButton>
     </DialogActions>
 </MudDialog>
 
-@code {
+@code 
+{
     [CascadingParameter] MudDialogInstance MudDialog { get; set; } = null!;
 
     [Parameter] public string Content { get; set; } = string.Empty;
+
+    [Parameter] public Dictionary<bool, string> SelectButtonDialog { get; set; } = new Dictionary<bool, string> { { true, "Yes" }, { false, "No" } };
 
     void Submit() => MudDialog.Close(DialogResult.Ok(true));
