+++ TestTask.MudBlazors/Pages/StartPage.razor
@@ -2,7 +2,7 @@
 
 @using System.Security.Claims
 @using TestTask.MudBlazors.Authenticate
-@using TestTask.MudBlazors.Pages.Table.Dialog
+@using TestTask.MudBlazors.Dialog
 
 @inject NavigationManager Navigation
