+++ TestTask.MudBlazors/Pages/Table/TablePage.razor
@@ -1,7 +1,8 @@
 ﻿@page "/table/"
 
  @using Table             
- @using TestTask.MudBlazors.Pages.Table.Dialog
+ @using TestTask.MudBlazors.Dialog
+ @using TestTask.MudBlazors.Dialog.ItemTable
  @using TestTask.MudBlazors.Pages.Table.Model
  @using TestTask.MudBlazors.Model
  @using TestTask.MudBlazors.Pages.Export
